# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMANTIC - Change & Configuration API",
        "description": "API untuk CMDB, Change Management, Maintenance & Patch Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/patch-catalogs": {
            "get": {
                "tags": [
                    "Patch Catalog"
                ],
                "summary": "Daftar patch catalog",
                "description": "GET /api/v1/patch-catalogs",
                "operationId": "67a7fb8b523779edee104d4ee916c484",
                "responses": {
                    "200": {
                        "description": "List patch catalog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "vendor": {
                                                "type": "string",
                                                "example": "Microsoft"
                                            },
                                            "patch_code": {
                                                "type": "string",
                                                "example": "KB5030219"
                                            },
                                            "severity": {
                                                "type": "string",
                                                "example": "critical"
                                            },
                                            "notes": {
                                                "type": "string",
                                                "example": "Security cumulative update November 2025"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Patch Catalog"
                ],
                "summary": "Menambahkan patch catalog baru",
                "description": "POST /api/v1/patch-catalogs",
                "operationId": "30bd436f893d22e1a775621f39d631e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendor",
                                    "patch_code",
                                    "severity"
                                ],
                                "properties": {
                                    "vendor": {
                                        "type": "string",
                                        "example": "Microsoft"
                                    },
                                    "patch_code": {
                                        "type": "string",
                                        "example": "KB5030219"
                                    },
                                    "severity": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "critical"
                                        ],
                                        "example": "critical"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Security cumulative update November 2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patch catalog berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "vendor": {
                                            "type": "string",
                                            "example": "Microsoft"
                                        },
                                        "patch_code": {
                                            "type": "string",
                                            "example": "KB5030219"
                                        },
                                        "severity": {
                                            "type": "string",
                                            "example": "critical"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/patch-catalogs/{id}": {
            "get": {
                "tags": [
                    "Patch Catalog"
                ],
                "summary": "Detail satu patch catalog",
                "description": "GET /api/v1/patch-catalogs/{patch_catalog}",
                "operationId": "8ee43e357efb2b0edbf433d407f8d712",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID patch catalog",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail patch catalog",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "vendor": {
                                            "type": "string",
                                            "example": "Microsoft"
                                        },
                                        "patch_code": {
                                            "type": "string",
                                            "example": "KB5030219"
                                        },
                                        "severity": {
                                            "type": "string",
                                            "example": "critical"
                                        },
                                        "notes": {
                                            "type": "string",
                                            "example": "Security cumulative update November 2025"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patch catalog tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}